LayoutTests/media/modern-media-controls/media-controls/media-controls-constructor-expected.txt

Testing the MediaControls constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.element.localName is "div"
PASS mediaControls.element.className is "media-controls"

PASS mediaControls.width is 300
PASS mediaControls.height is 150
PASS mediaControls.layoutTraits is LayoutTraits.macOS
PASS mediaControls.showsStartButton is false

PASS mediaControls.startButton instanceof StartButton is true
PASS mediaControls.playPauseButton instanceof PlayPauseButton is true
PASS mediaControls.skipBackButton instanceof SkipBackButton is true
PASS mediaControls.airplayButton instanceof AirplayButton is true
PASS mediaControls.pipButton instanceof PiPButton is true
PASS mediaControls.fullscreenButton instanceof FullscreenButton is true
PASS mediaControls.statusLabel instanceof StatusLabel is true
PASS mediaControls.timeControl instanceof TimeControl is true
PASS mediaControls.airplayPlacard instanceof AirplayPlacard is true
PASS mediaControls.invalidPlacard instanceof InvalidPlacard is true
PASS mediaControls.pipPlacard instanceof PiPPlacard is true

PASS mediaControls.controlsBar instanceof ControlsBar is true
PASS mediaControls.controlsBar.element.localName is "div"
PASS mediaControls.controlsBar.element.className is "controls-bar"

PASS mediaControlsWithArguments.width is 600
PASS mediaControlsWithArguments.height is 300

PASS successfullyParsed is true

TEST COMPLETE
